Add mocked Ads recommendations endpoint.
